seperation of passes
