+++ ToolKit/BloomPass.h
@@ -1,0 +1,43 @@
+#pragma once
+
+#include "Pass.h"
+#include "FullQuadPass.h"
+
+namespace ToolKit
+{
+
+  struct BloomPassParams
+  {
+    FramebufferPtr FrameBuffer = nullptr;
+    int iterationCount         = 6;
+    float minThreshold = 1.0f, intensity = 1.0f;
+  };
+
+  class TK_API BloomPass : public Pass
+  {
+   public:
+    BloomPass();
+    explicit BloomPass(const BloomPassParams& params);
+    ~BloomPass();
+
+    void Render() override;
+    void PreRender() override;
+    void PostRender() override;
+
+   public:
+    BloomPassParams m_params;
+
+   private:
+    // Iteration Count + 1 number of textures & framebuffers
+    std::vector<RenderTargetPtr> m_tempTextures;
+    std::vector<FramebufferPtr> m_tempFrameBuffers;
+    FullQuadPassPtr m_pass       = nullptr;
+    ShaderPtr m_downsampleShader = nullptr;
+    ShaderPtr m_upsampleShader   = nullptr;
+
+    bool m_invalidRenderParams   = false;
+  };
+
+  typedef std::shared_ptr<BloomPass> BloomPassPtr;
+
+} // namespace ToolKit
