+++ ToolKit/GammaPass.cpp
@@ -1,0 +1,34 @@
+#include "GammaPass.h"
+
+#include "Shader.h"
+#include "ShaderReflectionCache.h"
+#include "ToolKit.h"
+
+#include "DebugNew.h"
+
+namespace ToolKit
+{
+
+  GammaPass::GammaPass() : PostProcessPass()
+  {
+    m_postProcessShader = GetShaderManager()->Create<Shader>(
+        ShaderPath("gammaFrag.shader", true));
+  }
+
+  GammaPass::GammaPass(const GammaPassParams& params) : GammaPass()
+  {
+    m_params = params;
+  }
+
+  GammaPass::~GammaPass() {}
+
+  void GammaPass::PreRender()
+  {
+    PostProcessPass::m_params.FrameBuffer = m_params.FrameBuffer;
+    PostProcessPass::PreRender();
+
+    m_postProcessShader->SetShaderParameter("Gamma",
+                                            ParameterVariant(m_params.Gamma));
+  }
+
+} // namespace ToolKit
