+++ ToolKit/StencilPass.cpp
@@ -1,0 +1,83 @@
+#include "StencilPass.h"
+
+#include "Material.h"
+#include "Mesh.h"
+#include "ToolKit.h"
+
+namespace ToolKit
+{
+
+  StencilRenderPass::StencilRenderPass()
+  {
+    // Init sub pass.
+    m_copyStencilSubPass = std::make_shared<FullQuadPass>();
+    m_copyStencilSubPass->m_params.FragmentShader =
+        GetShaderManager()->Create<Shader>(
+            ShaderPath("unlitFrag.shader", true));
+
+    m_solidOverrideMaterial =
+        GetMaterialManager()->GetCopyOfUnlitColorMaterial();
+  }
+
+  StencilRenderPass::StencilRenderPass(const StencilRenderPassParams& params)
+      : StencilRenderPass()
+  {
+    m_params = params;
+  }
+
+  StencilRenderPass::~StencilRenderPass()
+  {
+    m_frameBuffer           = nullptr;
+    m_solidOverrideMaterial = nullptr;
+    m_copyStencilSubPass    = nullptr;
+  }
+
+  void StencilRenderPass::Render()
+  {
+    Renderer* renderer      = GetRenderer();
+    renderer->m_overrideMat = m_solidOverrideMaterial;
+
+    // Stencil pass.
+    renderer->SetStencilOperation(StencilOperation::AllowAllPixels);
+    renderer->ColorMask(false, false, false, false);
+
+    for (Entity* ntt : m_params.DrawList)
+    {
+      renderer->Render(ntt, m_params.Camera);
+    }
+
+    // Copy pass.
+    renderer->ColorMask(true, true, true, true);
+    renderer->SetStencilOperation(StencilOperation::AllowPixelsFailingStencil);
+
+    RenderSubPass(m_copyStencilSubPass);
+
+    renderer->SetStencilOperation(StencilOperation::None);
+  }
+
+  void StencilRenderPass::PreRender()
+  {
+    Pass::PreRender();
+
+    m_frameBuffer = std::make_shared<Framebuffer>();
+
+    FramebufferSettings settings;
+    settings.depthStencil    = true;
+    settings.useDefaultDepth = true;
+    settings.width           = m_params.OutputTarget->m_width;
+    settings.height          = m_params.OutputTarget->m_height;
+
+    m_frameBuffer->Init(settings);
+    m_frameBuffer->SetAttachment(Framebuffer::Attachment::ColorAttachment0,
+                                 m_params.OutputTarget);
+
+    m_copyStencilSubPass->m_params.FrameBuffer = m_frameBuffer;
+
+    Renderer* renderer                         = GetRenderer();
+    renderer->SetFramebuffer(m_frameBuffer, true, Vec4(0.0f));
+    renderer->SetCameraLens(m_params.Camera);
+  }
+
+  void StencilRenderPass::PostRender() { Pass::PostRender(); }
+
+} // namespace ToolKit
