+++ ToolKit/ToolKit.vcxproj
@@ -14,15 +14,23 @@
     <ClCompile Include="Animation.cpp" />
     <ClCompile Include="Audio.cpp" />
     <ClCompile Include="BinPack2D.cpp" />
+    <ClCompile Include="BloomPass.cpp" />
     <ClCompile Include="Canvas.cpp" />
     <ClCompile Include="Component.cpp" />
     <ClCompile Include="Camera.cpp" />
+    <ClCompile Include="CubemapPass.cpp" />
     <ClCompile Include="DataTexture.cpp" />
+    <ClCompile Include="DeferredPass.cpp" />
     <ClCompile Include="DirectionComponent.cpp" />
+    <ClCompile Include="DofPass.cpp" />
     <ClCompile Include="Drawable.cpp" />
     <ClCompile Include="Entity.cpp" />
     <ClCompile Include="FileManager.cpp" />
+    <ClCompile Include="ForwardPass.cpp" />
     <ClCompile Include="Framebuffer.cpp" />
+    <ClCompile Include="FullQuadPass.cpp" />
+    <ClCompile Include="FxaaPass.cpp" />
+    <ClCompile Include="GammaPass.cpp" />
     <ClCompile Include="GBufferPass.cpp" />
     <ClCompile Include="Light.cpp" />
     <ClCompile Include="Logger.cpp" />
@@ -30,6 +38,7 @@
     <ClCompile Include="MathUtil.cpp" />
     <ClCompile Include="Mesh.cpp" />
     <ClCompile Include="Node.cpp" />
+    <ClCompile Include="OutlinePass.cpp" />
     <ClCompile Include="ParameterBlock.cpp" />
     <ClCompile Include="Pass.cpp" />
     <ClCompile Include="PluginManager.cpp" />
@@ -45,16 +54,20 @@
     <ClCompile Include="Scene.cpp" />
     <ClCompile Include="SceneRenderer.cpp" />
     <ClCompile Include="Shader.cpp" />
+    <ClCompile Include="ShadowPass.cpp" />
     <ClCompile Include="Skeleton.cpp" />
     <ClCompile Include="Sky.cpp" />
     <ClCompile Include="SpriteSheet.cpp" />
+    <ClCompile Include="SsaoPass.cpp" />
     <ClCompile Include="StateMachine.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="StencilPass.cpp" />
     <ClCompile Include="Surface.cpp" />
     <ClCompile Include="Texture.cpp" />
+    <ClCompile Include="ToneMapPass.cpp" />
     <ClCompile Include="ToolKit.cpp" />
     <ClCompile Include="UIManager.cpp" />
     <ClCompile Include="Util.cpp" />
@@ -64,18 +77,27 @@
     <ClInclude Include="Animation.h" />
     <ClInclude Include="Audio.h" />
     <ClInclude Include="BinPack2D.h" />
+    <ClInclude Include="BloomPass.h" />
     <ClInclude Include="Canvas.h" />
     <ClInclude Include="Component.h" />
+    <ClInclude Include="CubemapPass.h" />
     <ClInclude Include="DataTexture.h" />
     <ClInclude Include="DebugNew.h" />
     <ClInclude Include="Camera.h" />
+    <ClInclude Include="DeferredPass.h" />
     <ClInclude Include="DirectionComponent.h" />
+    <ClInclude Include="DofPass.h" />
     <ClInclude Include="Drawable.h" />
     <ClInclude Include="Entity.h" />
     <ClInclude Include="Events.h" />
     <ClInclude Include="FileManager.h" />
+    <ClInclude Include="ForwardPass.h" />
     <ClInclude Include="Framebuffer.h" />
+    <ClInclude Include="FullQuadPass.h" />
+    <ClInclude Include="FxaaPass.h" />
+    <ClInclude Include="GammaPass.h" />
     <ClInclude Include="GBufferPass.h" />
+    <ClInclude Include="OutlinePass.h" />
     <ClInclude Include="Pass.h" />
     <ClInclude Include="PostProcessPass.h" />
     <ClInclude Include="RenderSystem.h" />
@@ -100,13 +122,17 @@
     <ClInclude Include="Scene.h" />
     <ClInclude Include="Serialize.h" />
     <ClInclude Include="Shader.h" />
+    <ClInclude Include="ShadowPass.h" />
     <ClInclude Include="Skeleton.h" />
     <ClInclude Include="Sky.h" />
     <ClInclude Include="SpriteSheet.h" />
+    <ClInclude Include="SsaoPass.h" />
     <ClInclude Include="StateMachine.h" />
     <ClInclude Include="stdafx.h" />
+    <ClInclude Include="StencilPass.h" />
     <ClInclude Include="Surface.h" />
     <ClInclude Include="Texture.h" />
+    <ClInclude Include="ToneMapPass.h" />
     <ClInclude Include="ToolKit.h" />
     <ClInclude Include="Types.h" />
     <ClInclude Include="UIManager.h" />
